refactor(model): narrow byte_to_cp to uint16_t, drop dead min_cp entries (#26)
byte_to_cp values top out at 323, so uint16_t halves the table; the
min_cp entries for byte_len 0/1 were unreachable (ASCII returns before
the semantic check), so index by byte_len - 2 over the three real
minimums. Behavior unchanged, existing roundtrip/overlong tests cover.

+++ src/model/tok_unicode.h
@@ -27,7 +27,8 @@ bool uc_is_whitespace_cp(uint32_t cp);
 #define UC_BYTES_UNICODE_REV_SIZE 324
 
 typedef struct {
-    uint32_t byte_to_cp[256];
+    /* Max mapped codepoint is 323, so uint16_t suffices. */
+    uint16_t byte_to_cp[256];
     /* UINT16_MAX = no byte maps to this codepoint. */
     uint16_t cp_to_byte[UC_BYTES_UNICODE_REV_SIZE];
 } uc_bytes_unicode_t;
